Update kalman-filter.md
In the first pseudocode of the Kalman Filter the Kalman Gain is directly added to the current state instead of being multiplied by the residual. This means that the residual variable is unused and that the Kalman Gain isn't actually determining how much weight to assign to both the measurements and prediction.

This has been added in line 51 of the Pull Request.